requirements.txt
# requirements.txt (Resolved dependencies)

# Core AI Agent Framework and Tools
crewai
crewai-tools

# Langchain Core and Integrations - updated to compatible versions
langchain>=0.3.0
langchain-community>=0.2.0
langchain-core>=0.3.52
langchain-google-genai==2.1.3
langchain-groq>=0.1.5
langchain-google-community>=1.0.2

# Web framework for the UI
streamlit

# Utilities and Dependencies
pydantic>=2.7.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
requests>=2.31.0
tenacity>=8.2.3
nest_asyncio
langchain-google-community[places] # New Tool Dependencies
google-api-python-client>=2.108.0
# openweathermap-api-wrapper  # Removed - doesn't exist as separate package
wikipedia
yelpapi
playwright
pyowm